csi3335_baseball_app/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
PyMySQL==1.1.1
pandas==2.2.2
numpy==1.26.4
Werkzeug==3.0.3
email-validator==2.1.1
